поправляет задание "Больше деталей"
