+++ js/isCount.js
@@ -1,0 +1,11 @@
+const isCount = () => {
+  let count = 0;
+
+  return function () {
+    return count++;
+  };
+};
+
+export {
+  isCount
+};
