+++ js/main.js
@@ -1,8 +1,8 @@
-import { createArrayPhotos } from './createObject.js';
+import { createObjectPhotos } from './createObject.js';
 
 const COUNT_COMMENTS = 25;
 
-const arrayPhotos = Array.from({ length: COUNT_COMMENTS }, createArrayPhotos);
+const arrayPhotos = Array.from({ length: COUNT_COMMENTS }, createObjectPhotos);
 
 // console.table(arrayPhotos);
